コンポーネント単位でファイルを分割
