+++ src/components/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Game from './Game';
+import '../index.css';
+
+class App extends React.Component {
+    render() {
+        return (
+            <Game />
+        );
+    }
+}
+
+export default App;
